Apply reusable page title card to pages
