+++ movie_shape/lib/views/watchlist/watchlist_view.dart
@@ -3,6 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:movie_shape/models/film.dart';
 import 'package:movie_shape/repository/film_repo_implemented.dart';
 import 'package:movie_shape/reusable/list/list_card.dart';
+import 'package:movie_shape/reusable/page_title_card.dart';
 import 'package:movie_shape/views/pages/film_detail_page.dart';
 import 'package:movie_shape/views/watchlist/state/watchlist_bloc/watchlist_bloc.dart';
 import 'package:movie_shape/views/widgets/searchbar.dart';
@@ -24,13 +25,24 @@ class WatchlistView extends StatelessWidget {
 
     return Column(
       children: [
-        CustomSearchBar(
-          onSearchResults: (searchQuery) {
-            context.read<WatchlistBloc>().add(SearchWatchlist(searchQuery));
-          },
-          onClear: () =>
-              {context.read<WatchlistBloc>().add(ClearWatchSearch())},
-          hintText: "Search your Watch List...",
+        // Title card
+        Padding(
+          padding:
+              const EdgeInsets.only(left: 20, top: 20, right: 20, bottom: 14),
+          child: PageTitleCard(title: 'Favourites'),
+        ),
+        // Search bar
+        Padding(
+          padding:
+              const EdgeInsets.only(left: 20, top: 14, right: 20, bottom: 28),
+          child: CustomSearchBar(
+            onSearchResults: (searchQuery) {
+              context.read<WatchlistBloc>().add(SearchWatchlist(searchQuery));
+            },
+            onClear: () =>
+                {context.read<WatchlistBloc>().add(ClearWatchSearch())},
+            hintText: "Search your Watch List...",
+          ),
         ),
         Expanded(
           child: BlocBuilder<WatchlistBloc, WatchlistState>(
